move pretty-printing code
